+++ data-raw/sbcb.R
@@ -1,0 +1,4 @@
+library(bcbioSmallRna)
+
+sbcb = loadSmallRnaRun("inst/extra/bcbio", interestingGroups = "country")
+save(sbcb, file = "data/sbcb.rda", compress = "xz")