+++ NEWS.md
@@ -1,6 +1,7 @@
 # 0.0.1
 
 ## Fix
+* Add plots to the template and remove it from the function
 * Fix bcbioSmallSizeDist to avoid duplication of columns when getting adapter information
 
 ## Feature